python-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sievedata==0.0.1.1.6
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
supabase
python-dotenv 